refactor: menu 컨텍스트 패키지 분리
